new _servo_pa algorithm.
Instead of looking for a +-5% window, just step toward the
target mixer current until we pass it (or hit a limit),
then set PA to the value that minimized the error.

Also adjust the PA step size to avoid aliasing due to
floating point rounding and PA scale voltage quantization.