bitwise OR simulate param instead of override.
Previously, the __init__ simulate parameter would
replace the config if set.  This made it tricky to use
properly, and in practice it wasn't used anywhere.

Now the simulate parameter is ORed with the config,
which means it can be used to exclude certain parts
of the system as needed.  For instance, power_down can
ignore Load/IFswitch while still using Instrument.